+++ src/main/java/domain/menu/MenuRepository.java
@@ -1,9 +1,11 @@
-package domain;
+package domain.menu;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+
+import domain.Category;
 
 public class MenuRepository {
     private static final List<Menu> menus = new ArrayList<>();
