+++ src/test/java/domain/menu/MenuRepositoryTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import static org.assertj.core.api.Assertions.assertThat;
