+++ src/test/java/domain/menu/MenuTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.menu;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import domain.Category;
 
 class MenuTest {
     private static final Menu CHICKEN = new Menu(1, "양념치킨", Category.CHICKEN, 16_000);
