+++ src/test/java/domain/table/TableTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.table;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
